requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
pgvector
sentence-transformers
groq
sqlglot
python-dotenv
streamlit